requirements.txt
certifi==2025.1.31
contourpy==1.3.2
cycler==0.12.1
fonttools==4.57.0
ImageTopoDec==1.0.7
kiwisolver==1.4.8
laspy==2.5.4
matplotlib==3.10.3
numpy==2.2.5
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
pyshp==2.3.1
python-dateutil==2.9.0.post0
six==1.17.0
